requirements.txt
requests
pyyaml
docker
